Improve zsh history for my workflow
When working in multiple terminals on discrete tasks, it's annoying to
have the history from one show up in another right away. It throws off
being able to press the up arrow to return to steps in one of the
discrete tasks. Also, with these discrete tasks, it makes more sense to
save the commands grouped per terminal rather than at the time they are
entered, which results in the commands for two tasks being intermingled.